work on weak residual functions

+++ src/symbolic_utils.h
@@ -12,7 +12,7 @@ create_symbolic_system(const std::vector<std::string> &f);
 std::vector<Expression> create_symbolic_vars(const std::string &base_name,
                                              size_t count);
 
-std::vector<SymEngine::Expression> create_all_symbolic_inputs(int J, int D);
+std::vector<SymEngine::Expression> create_all_symbolic_inputs(size_t J, size_t D);
 
 std::vector<std::vector<Expression> >
 compute_jacobian(const std::vector<Expression> &system,
@@ -21,7 +21,7 @@ compute_jacobian(const std::vector<Expression> &system,
 vec_basic expressions_to_vec_basic(const std::vector<Expression> &exprs);
 
 std::vector<LambdaRealDoubleVisitor>
-build_symbolic_system(const std::vector<Expression> &dx, int D, int J);
+build_symbolic_system(const std::vector<Expression> &dx, size_t D, size_t J);
 
 std::vector<std::vector<std::vector<Expression> > >
 compute_jacobian(const std::vector<std::vector<Expression> > &matrix,
